Separate domain from I/O.
Domain objects were coupled with I/O in form of
puts statements. This introduces brittleness or
resistence to change. Now we can create seperate
processors if we want to say send the rules via an
HTTP post etc. Also code coupled with I/O is hard to
test. This problem would have surfaced earlier if we had
unit tests.